Passing the test methods users to the corresponding class
